Rework of the grammar and now the tokenizer can handle number tokens
I'm getting closer to having a pretty solid grammar. The idea at this point is that we now have a concept called "Macros" (previously this was "Variables"). Unlike before, there is not difference between a macro and a macro function, and there is not longer a macro set. Instead, there is one type of macro, that can optionally be defined with a set of parameters, and when using a macro, you can access any of its members or values (in the appropriate situation - e.g. objects have members, arrays have values).

Continuing on with more prototyping and testing. Also, I'm starting to think about and prototype the parser.