remove the Zig lib from this repo -- this will only have specification
I've decided that this repo really should only contain the ZOMB file
format specification. I originally had the Zig library I was prototyping
in this same repo...just because I was prototyping.

Now that I'm getting closer to finalizing the file format, I need to
start cleaning things up. I will be keeping the entire prototyping Git
history of the Zig library in this repo -- it's kind of cool to look
back on the process of putting this all together even though it's an
embarrassing mess -- go ahead and judge me if you want, I don't give a
shit ;).